refactor memory storage
