+++ common/storage/storage.go
@@ -1,0 +1,11 @@
+package storage
+
+import (
+	"github.com/pastelnetwork/gonode/common/storage/definition"
+	"github.com/pastelnetwork/gonode/common/storage/memory"
+)
+
+//NewKeyValue return new instance key value storage
+func NewKeyValue() definition.KeyValue {
+	return memory.NewMemory()
+}
