+++ supernode/cmd/app.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pastelnetwork/gonode/common/errors"
 	"github.com/pastelnetwork/gonode/common/log"
 	"github.com/pastelnetwork/gonode/common/log/hooks"
+	"github.com/pastelnetwork/gonode/common/storage"
 	"github.com/pastelnetwork/gonode/common/sys"
 	"github.com/pastelnetwork/gonode/common/version"
 	"github.com/pastelnetwork/gonode/pastel-client"
@@ -17,7 +18,6 @@ import (
 	"github.com/pastelnetwork/gonode/supernode/node/grpc/services/supernode"
 	"github.com/pastelnetwork/gonode/supernode/node/grpc/services/walletnode"
 	"github.com/pastelnetwork/gonode/supernode/services/artworkregister"
-	"github.com/pastelnetwork/gonode/supernode/storage/memory"
 )
 
 const (
@@ -90,7 +90,7 @@ func runApp(ctx context.Context, config *configs.Config) error {
 	// entities
 	pastelClient := pastel.NewClient(config.Pastel)
 	nodeClient := grpc.NewClient()
-	db := memory.NewKeyValue()
+	db := storage.NewKeyValue()
 
 	// business logic services
 	artworkRegister := artworkregister.NewService(config.ArtworkRegister, db, pastelClient, nodeClient)
